feat: improve no key flow, moved check to checkout page
