+++ src/renderer/marketplace/bank-accounts/checkout/checkout-container.jsx
@@ -2,8 +2,10 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { push } from 'connected-react-router';
 import { withStyles } from '@material-ui/core/styles';
+import { BigNumber } from 'bignumber.js';
 import config from 'common/config';
 import EthUnits from 'common/utils/eth-units';
+import { getCryptoValue } from '../../../common/price-utils';
 import { getLocale } from 'common/locale/selectors';
 import { getFiatCurrency } from 'common/fiatCurrency/selectors';
 import { getTokens } from 'common/wallet-tokens/selectors';
@@ -54,31 +56,42 @@ class BankAccountsCheckoutContainer extends MarketplaceBankAccountsComponent {
 		};
 	}
 
-	onBackClick = () => this.props.dispatch(push(this.cancelRoute()));
+	priceInKEY = priceUSD => new BigNumber(priceUSD).dividedBy(this.props.keyRate);
+
+	keyAvailable = () => new BigNumber(this.props.cryptoValue);
 
 	onStartClick = async () => {
 		const { jurisdiction, templateId, vendorId, vendor } = this.props;
 		const { region } = jurisdiction.data;
 
-		// TODO: get URLs and vendor name from the RP store
-		this.props.dispatch(
-			kycOperations.startCurrentApplicationOperation(
-				vendorId,
-				templateId,
-				this.payRoute(),
-				this.cancelRoute(),
-				`Bank Account in ${region}`,
-				`You are about to begin the application process for a bank account in ${region}.
-				Please double check your required documents are Certified True or Notarized where
-				necessary. Failure to do so will result in delays in the process. You may also be
-				asked to provide more information by the service provider`,
-				'conducting KYC',
-				vendor.name,
-				vendor.privacyPolicy,
-				vendor.termsOfService
-			)
-		);
+		const keyPrice = this.priceInKEY(jurisdiction.price);
+		const keyAvailable = this.keyAvailable();
+		const transactionNoKeyError = `/main/transaction-no-key-error/${keyPrice}`;
+
+		if (keyPrice.gt(keyAvailable)) {
+			return this.props.dispatch(push(transactionNoKeyError));
+		} else {
+			this.props.dispatch(
+				kycOperations.startCurrentApplicationOperation(
+					vendorId,
+					templateId,
+					this.payRoute(),
+					this.cancelRoute(),
+					`Bank Account in ${region}`,
+					`You are about to begin the application process for a bank account in ${region}.
+					Please double check your required documents are Certified True or Notarized where
+					necessary. Failure to do so will result in delays in the process. You may also be
+					asked to provide more information by the service provider`,
+					'conducting KYC',
+					vendor.name,
+					vendor.privacyPolicy,
+					vendor.termsOfService
+				)
+			);
+		}
 	};
+
+	onBackClick = () => this.props.dispatch(push(this.cancelRoute()));
 
 	render() {
 		const { jurisdiction, countryCode } = this.props;
@@ -99,12 +112,9 @@ class BankAccountsCheckoutContainer extends MarketplaceBankAccountsComponent {
 				{...this.getPaymentParameters()}
 				price={jurisdiction.price}
 				options={checkoutOptions}
-				initialDocsText={`You will be required to provide a few basic information about yourself like full name and email.
-					This will be done through SelfKey ID Wallet.`}
-				kycProcessText={`You will undergo a standard KYC process and our team will get in touch with you to make sure we
-					have all the information needed.`}
-				getFinalDocsText={`Once the account opening process is done you will receive all the relevant documents, access codes
-					in persion/via courier or on your email.`}
+				initialDocsText={`You will be required to provide a few basic information about yourself like full name and email. This will be done through SelfKey ID Wallet.`}
+				kycProcessText={`You will undergo a standard KYC process and our team will get in touch with you to make sure we have all the information needed.`}
+				getFinalDocsText={`Once the account opening process is done you will receive all the relevant documents, access codes in persion/via courier or on your email.`}
 				onBackClick={this.onBackClick}
 				onStartClick={this.onStartClick}
 				startButtonText={'Start Application'}
@@ -117,6 +127,7 @@ const mapStateToProps = (state, props) => {
 	const { accountCode, vendorId, countryCode, templateId } = props.match.params;
 	const authenticated = true;
 	const identity = identitySelectors.selectIdentity(state);
+	const primaryToken = { ...props, cryptoCurrency: config.constants.primaryToken };
 	return {
 		countryCode,
 		templateId,
@@ -131,6 +142,7 @@ const mapStateToProps = (state, props) => {
 		...getFiatCurrency(state),
 		...ethGasStationInfoSelectors.getEthGasStationInfo(state),
 		tokens: getTokens(state).splice(1), // remove ETH
+		cryptoValue: getCryptoValue(state, primaryToken),
 		address: getWallet(state).address,
 		keyRate: pricesSelectors.getRate(state, 'KEY', 'USD'),
 		ethRate: pricesSelectors.getRate(state, 'ETH', 'USD'),
