+++ src/renderer/marketplace/incorporation/checkout/incorporations-checkout-container.jsx
@@ -2,8 +2,10 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { push } from 'connected-react-router';
 import { withStyles } from '@material-ui/core/styles';
+import { BigNumber } from 'bignumber.js';
 import config from 'common/config';
 import EthUnits from 'common/utils/eth-units';
+import { getCryptoValue } from '../../../common/price-utils';
 import { getLocale } from 'common/locale/selectors';
 import { getFiatCurrency } from 'common/fiatCurrency/selectors';
 import { getTokens } from 'common/wallet-tokens/selectors';
@@ -22,9 +24,7 @@ const FIXED_GAS_LIMIT_PRICE = 21000;
 class IncorporationsCheckoutContainer extends MarketplaceIncorporationsComponent {
 	async componentDidMount() {
 		this.props.dispatch(ethGasStationInfoOperations.loadData());
-
 		await this.loadRelyingParty({ rp: this.props.vendorId, authenticated: true });
-
 		this.checkIfUserCanIncorporate();
 	}
 
@@ -55,40 +55,53 @@ class IncorporationsCheckoutContainer extends MarketplaceIncorporationsComponent
 		};
 	}
 
-	onBackClick = () => this.props.dispatch(push(this.cancelRoute()));
+	priceInKEY = priceUSD => new BigNumber(priceUSD).dividedBy(this.props.keyRate);
+
+	keyAvailable = () => new BigNumber(this.props.cryptoValue);
 
 	onStartClick = async () => {
 		const { program, vendorId, templateId, vendor } = this.props;
 		const { region } = program.data;
 
-		this.props.dispatch(
-			kycOperations.startCurrentApplicationOperation(
-				vendorId,
-				templateId,
-				this.payRoute(),
-				this.cancelRoute(),
-				`Incorporate in ${region}`,
-				`You are about to begin the incorporation process in ${region}. Please double check your
-				required documents are Certified True or Notarized where necessary. Failure to do so
-				will result in delays in the incorporation process. You may also be asked to provide
-				more information by the service provider`,
-				'conducting KYC',
-				vendor.name,
-				vendor.privacyPolicy,
-				vendor.termsOfService
-			)
-		);
+		const keyPrice = this.priceInKEY(program.price);
+		const keyAvailable = this.keyAvailable();
+		const transactionNoKeyError = `/main/transaction-no-key-error/${keyPrice}`;
+
+		if (keyPrice.gt(keyAvailable)) {
+			return this.props.dispatch(push(transactionNoKeyError));
+		} else {
+			this.props.dispatch(
+				kycOperations.startCurrentApplicationOperation(
+					vendorId,
+					templateId,
+					this.payRoute(),
+					this.cancelRoute(),
+					`Incorporate in ${region}`,
+					`You are about to begin the incorporation process in ${region}. Please double check your
+					required documents are Certified True or Notarized where necessary. Failure to do so
+					will result in delays in the incorporation process. You may also be asked to provide
+					more information by the service provider`,
+					'conducting KYC',
+					vendor.name,
+					vendor.privacyPolicy,
+					vendor.termsOfService
+				)
+			);
+		}
 	};
+
+	onBackClick = () => this.props.dispatch(push(this.cancelRoute()));
 
 	render() {
 		const { program, countryCode } = this.props;
+		const { region } = program.data;
 		const description = program.data.walletDescription
 			? program.data.walletDescription
 			: program.data.servicesDescription;
 
 		return (
 			<PaymentCheckout
-				title={`Pay Incorporation Fee: ${program.data.region}`}
+				title={`Pay Incorporation Fee: ${region}`}
 				description={description}
 				timeToForm={program.data.timeToFormWeeks}
 				program={program}
@@ -99,15 +112,9 @@ class IncorporationsCheckoutContainer extends MarketplaceIncorporationsComponent
 				onBackClick={this.onBackClick}
 				onStartClick={this.onStartClick}
 				startButtonText={'Start Application'}
-				initialDocsText={
-					'You will be required to provide a few basic information about yourself like full name and email. This will be done through SelfKey ID Wallet.'
-				}
-				kycProcessText={
-					'You will undergo a standard KYC process and our team will get in touch with you to make sure we have all the information needed.'
-				}
-				getFinalDocsText={
-					'Once the incorporations process is done you will receive all the relevant documents, for your new company, on your email.'
-				}
+				initialDocsText={`You will be required to provide a few basic information about yourself like full name and email. This will be done through SelfKey ID Wallet.`}
+				kycProcessText={`You will undergo a standard KYC process and our team will get in touch with you to make sure we have all the information needed.`}
+				getFinalDocsText={`Once the Incorporation process is done you will receive all the relevant documents, for your new company, on your email.`}
 			/>
 		);
 	}
@@ -116,6 +123,7 @@ class IncorporationsCheckoutContainer extends MarketplaceIncorporationsComponent
 const mapStateToProps = (state, props) => {
 	const { companyCode, countryCode, templateId, vendorId } = props.match.params;
 	const authenticated = true;
+	const primaryToken = { ...props, cryptoCurrency: config.constants.primaryToken };
 	return {
 		countryCode,
 		templateId,
@@ -130,6 +138,7 @@ const mapStateToProps = (state, props) => {
 		...ethGasStationInfoSelectors.getEthGasStationInfo(state),
 		tokens: getTokens(state).splice(1), // remove ETH
 		address: getWallet(state).address,
+		cryptoValue: getCryptoValue(state, primaryToken),
 		keyRate: pricesSelectors.getRate(state, 'KEY', 'USD'),
 		ethRate: pricesSelectors.getRate(state, 'ETH', 'USD'),
 		cryptoCurrency: CRYPTOCURRENCY,
